Skillbox/Telegram_bot/Project_bot/requirements.txt
aiogram==3.2.0
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
bs4==0.0.1
peewee==3.17.6